# Optimize value_counts function for performance improvement with missing classes